requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
pydantic
pandas
openpyxl
pillow
boto3
requests
moto
python-multipart
python-dotenv
aiohttp